% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sim.R
\name{do_sim}
\alias{do_sim}
\title{Makes Formula}
\usage{
do_sim(model, dataframe)
}
\arguments{
\item{model}{A fitted mediator regression model}

\item{dataframe}{Data frame containing data to be analysed. The data frame has cases in rows and variables in columns.}
}
\value{
\item{dataframe }{dataframe}
}
\description{
Given a causal
}
\examples{
\dontrun{
# I don't want you to run this
}
}
\keyword{internal}
