% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_make_formula.R
\name{eval_make_formula}
\alias{eval_make_formula}
\title{Evaluates and Makes Formula for regression of exposure or risk factor or outcome on its parents in a causal Bayesian network directed acyclic graph.}
\usage{
eval_make_formula(dataframe, in_out, model_list, w)
}
\arguments{
\item{dataframe}{A wide format dataframe containing all the risk factors, confounders, exposures and outcomes within the causal DAG Bayesian network.}

\item{in_out}{A list of length 2. The first list contains the ``in_vars'' which is a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list conttains a list of a single name of exposure or risk factor or outcome in form of characters.}

\item{model_list}{list of a single name of exposure or risk factor or outcome in form of characters.}

\item{w}{Column of weights for case control matching listing in same order as patients in dataframe.}
}
\value{
\item{model_list }{model list}
}
\description{
Evaluates and Makes Formula for regression of exposure or risk factor or outcome on its parents in a causal Bayesian network directed acyclic graph. Given a causal Bayesian network, directed acyclic graph (DAG) where arrows representing
causal dependencies between confounders, risk factors/exposure and disease, together with a sensible probability distribution on
the graph that respects these causal dependencies. To consistently estimate causal effects that risk factors may have on each
other and on disease, we need to make a strong no unmeasured confounding assumption: that is common causes of nodes in the graph,
which may be causes of two risk factors or a cause of risk factor and disease, are also included as nodes in the graph.
Causal Bayesian networks have a local Markov property that the conditional probability distribution of any node Xj, given values
for the other variables in the network, only depends on the values $x_{pa}_{j}$ of the parent nodes.
}
\examples{
\dontrun{
# I don't want you to run this
}
in_vars = c("subeduc","moteduc","fatduc")
outvar = c("phys")
make_formula(in_vars,outvar)
}
\keyword{Regression}
\keyword{models}
