% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DAG.R
\name{make_DAG}
\alias{make_DAG}
\title{Creates a DAG for input into package Dagitty to identify adjustmentSets given exposure and outcome}
\usage{
make_DAG(
  in_outDAG,
  exposure,
  response,
  mediator,
  Splines_outlist_Var,
  splinesDefinedIn_in_outDAG,
  addCustomExposureAdjustmentSet = FALSE,
  customExposureAdjustmentSet,
  addCustomMediatorAdjustmentSet = FALSE,
  customMediatorAdjustmentSet
)
}
\arguments{
\item{in_outDAG}{A list of length 2. The first list contains a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list contains a list of a single name of exposure or risk factor or outcome in form of characters. See tutorial examples for examples.}

\item{exposure}{The name of the exposure column variable within dataframe in text format e.g. "phys".}

\item{response}{The name of the response column variable within dataframe in text format e.g. "case". The cases should be coded as 1 and the controls as 0.}

\item{mediator}{The name of the mediator column variables within dataframe in text format. There can be more than one mediator of interest. It can be a vector of mediators names within the dataframe e.g. c("subhtn","apob_apoa","whr").}

\item{Splines_outlist_Var}{A list defined of same size and order of variables as defined in in_outArg[[2]]. If splines are to be used for variables listed in in_outArg[[2]], then the splines should be defined in the same order as variables appear in in_outArg[[2]]. It is necessary to list variables in in_outArg[[2]] without splines if no spline is to be applied. Should be input as list() if no splines.}

\item{splinesDefinedIn_in_outDAG}{Logical TRUE or FALSE indicating whether the user has defined splines in the causal DAG, in_out, if TRUE. If FALSE and splines are defined in Splines_outlist_Var, then it is necessary for the package to populate the in_out DAG with splines listed in Splines_outlist_Var.}

\item{addCustomExposureAdjustmentSet}{Logical TRUE or FALSE indicating whether a customised interaction term is to be added to the each regression for the Exposure adjustment set. The interaction term can include splines. NOTE variables in addCustom are input by the User and the User should decide whether they are to be included in the adjustmentSet and whether these variables are not already present. The DAG will be fit without nodes in addCustom since causal DAG nodes will not recognise customised nodes e.g. with customed formula in the node name. The addCustom is added in additivly as adjustment set variables after the causal diagram has been defined and adjustment sets have been identified. Care should be taken when using addCustom not to add in variables eg. twice if already included as variables in the causal dag.}

\item{customExposureAdjustmentSet}{text containing the customised interaction term to be added to each regression for the Exposure adjustment set. The text should be enclosed in inverted commas. Splines can be included within the interactin terms. If there are more than one exposure and the custom is to differ per exposoure, custom can be defined in a vector of form c(), where each customised adjustment is in the same order as the variables defined in exposure if more than one expoosure.  See tutorial for examples. NOTE variables in addCustom are input by the User and the User should decide whether they are to be included in the adjustmentSet and whether these variables are not already present. The DAG will be fit without nodes in addCustom since causal DAG nodes will not recognise customised nodes e.g. with customed formula in the node name. The addCustom is added in additivly as adjustment set variables after the causal diagram has been defined and adjustment sets have been identified. Care should be taken when using addCustom not to add in variables eg. twice if already included as variables in the causal dag.}

\item{addCustomMediatorAdjustmentSet}{Logical TRUE or FALSE indicating whether a customised interaction term is to be added to the each regression for the Mediator adjustment set. The interaction term can include splines. NOTE variables in addCustom are input by the User and the User should decide whether they are to be included in the adjustmentSet and whether these variables are not already present. The DAG will be fit without nodes in addCustom since causal DAG nodes will not recognise customised nodes e.g. with customed formula in the node name. The addCustom is added in additivly as adjustment set variables after the causal diagram has been defined and adjustment sets have been identified. Care should be taken when using addCustom not to add in variables eg. twice if already included as variables in the causal dag.}

\item{customMediatorAdjustmentSet}{text containing the customised interaction term to be added to each regression for the Mediator adjustment set. The text should be enclosed in inverted commas. Splines can be included within the interactin terms.  If there are more than one mediatr and the custom is to differ per mediator, custom can be defined in a vector of form c(), where each customised adjustment is in the same order as the variables defined in mediator if more than one mediator. See tutorial for examples. NOTE variables in addCustom are input by the User and the User should decide whether they are to be included in the adjustmentSet and whether these variables are not already present. The DAG will be fit without nodes in addCustom since causal DAG nodes will not recognise customised nodes e.g. with customed formula in the node name. The addCustom is added in additivly as adjustment set variables after the causal diagram has been defined and adjustment sets have been identified. Care should be taken when using addCustom not to add in variables eg. twice if already included as variables in the causal dag.}
}
\value{
Returns DAG and adjustment set given exposure and outcome.
}
\description{
Creates a DAG for input into package Dagitty to identify adjustmentSets given exposure and outcome
}
\keyword{Attributable}
\keyword{Fraction}
\keyword{Population}
\keyword{Regression}
\keyword{models}
