% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_direct.R
\name{overall_direct}
\alias{overall_direct}
\title{Calculation of Population Attributable Fraction (PAF), with a decomposition of the total PAF into direct
and indirect components.}
\usage{
overall_direct(
  data_frame,
  exposure,
  mediator,
  response,
  mediator_model,
  response_model,
  response_model_2,
  weights
)
}
\arguments{
\item{data_frame}{Data frame containing the data. The data frame has cases in rows and variables in columns.}

\item{exposure}{The exposure name in the form of character string e.g. "phys" for physical exercise.}

\item{mediator}{The mediator name in the form of character string e.g. "whr" for waist-hip ratio.}

\item{response}{The outcome name in the form of character string e.g. "case" for a stroke case.}

\item{mediator_model}{A list containing each of the mediator regression models e.g. \code{mediator_model=list(model_list[[6]],model_list[[7]],model_list[[8]])}.}

\item{response_model}{is a regression model fitted for the outcome on all mediators together with all parents and confounders of the mediators in
a Markov causal Bayesian network DAG.}

\item{response_model_2}{is a regression model fitted for the outcome}

\item{weights}{A numeric}
}
\value{
\item{directPAF}{direct PAF}
}
\description{
Total PAF
}
\examples{
\dontrun{
# I don't want you to run this
}
}
\keyword{Regression}
\keyword{models}
