% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_vs_exposure.R
\name{response_vs_exposure}
\alias{response_vs_exposure}
\title{Checks if the causal DAG satisfies the Markov condition}
\usage{
response_vs_exposure(data, exposure = "phys", response = "case", in_out, w)
}
\arguments{
\item{data}{data stored as a dataframe with columns as variables and rows as cases and controls in case control study.}

\item{exposure}{exposure name in text format e.g. exposure="phys"}

\item{response}{response name in text format e.g. response="case"}

\item{in_out}{A list of length 2. The first list contains a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list conttains a list of a single name of exposure or risk factor or outcome in form of characters. See tutorial examples for examples.}

\item{w}{Column of weights for case control matching listing in same order as patients in data.}
}
\value{
\item{to_execute }{to_execute is in text format and represents the response_vs_exposure model to be evaluated including all covariates below and at the same level of the exposure in the causal Bayesian DAG.}
}
\description{
The functions checks if the Markov condition holds for the Directed Acyclic Graph (DAG) defined. Sometimes called the Markov assumption, is an assumption made in Bayesian probability theory, that every node in a Bayesian network is conditionally independent of its nondescendents, given its parents. In other words, it is assumed that a node has no bearing on nodes which do not descend from it. This is equivalent to stating that a node is conditionally independent of the entire network, given its Markov blanket. The related Causal Markov condition states that, conditional on the set of all its direct causes, a node is independent of all variables which are not direct causes or direct effects of that node.
}
\examples{
\dontrun{
# I don't want you to run this
}
in_vars = c("subeduc","moteduc","fatduc")
outvar = c("phys")
make_formula(in_vars,outvar)
}
\keyword{Acyclic}
\keyword{Attributable}
\keyword{Bayesian}
\keyword{Directed}
\keyword{Fraction}
\keyword{Graph}
\keyword{Markov}
\keyword{Population}
\keyword{models}
