% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strokedata.R
\docType{data}
\name{strokedata}
\alias{strokedata}
\title{Stroke data}
\format{
A data frame of random generated stroke data
}
\usage{
strokedata
}
\description{
Best data ever
}
\keyword{"data/maurice_data.rda",}
\keyword{#}
\keyword{2}
\keyword{2)}
\keyword{3}
\keyword{3.5.0}
\keyword{4.0.0}
\keyword{4.0.2.}
\keyword{<}
\keyword{=}
\keyword{Because}
\keyword{R}
\keyword{Rather,}
\keyword{The}
\keyword{able}
\keyword{and}
\keyword{be}
\keyword{can't}
\keyword{created}
\keyword{datasets}
\keyword{doesn't}
\keyword{file}
\keyword{files}
\keyword{files.}
\keyword{in}
\keyword{is}
\keyword{it's}
\keyword{load}
\keyword{matrix(rnorm(100),}
\keyword{maurice_data}
\keyword{mean/isn't}
\keyword{ncol=10)}
\keyword{others}
\keyword{otherwise}
\keyword{save(maurice_data,}
\keyword{saved}
\keyword{saving}
\keyword{telling}
\keyword{that}
\keyword{to}
\keyword{use}
\keyword{using}
\keyword{version}
\keyword{warning}
\keyword{when}
\keyword{won't}
\keyword{you}
\keyword{your}
