% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalPAFplot.R
\name{causalPAFplot}
\alias{causalPAFplot}
\title{Evaluates Total PAF, Direct PAF, Indirect PAF and Path Specific PAF for a user inputted number of bootstraps and integral simulations}
\usage{
causalPAFplot(
  dataframe,
  exposure = "phys",
  mediator = c("subhtn", "apob_apoa", "whr"),
  response = "case",
  mediator_model = list(model_list[[6]], model_list[[7]], model_list[[8]]),
  response_model_mediators = response_vs_mediator,
  response_model_exposure = response_vs_phys,
  in_out,
  model_list = list(),
  w,
  NumBootstrap,
  NumSimulation
)
}
\arguments{
\item{dataframe}{A wide format dataframe containing all the risk factors, confounders, exposures and outcomes within the causal DAG Bayesian network.}

\item{exposure}{The name of the exposure column variable within dataframe in text format e.g. "phys".}

\item{mediator}{The name of the mediator column variables within dataframe in text format. There can be more than one mediator of interest. It can be a vector of mediators names within the dataframe e.g. c("subhtn","apob_apoa","whr").}

\item{response}{The name of the response column variable within dataframe in text format e.g. "case". The cases should be coded as 1 and the controls as 0.}

\item{mediator_model}{A vector of mediator models fitted which must correspond to the same order of the mediator names in the variable ``mediator''.}

\item{response_model_mediators}{A model fitted for the response in a causal Bayesian network excluding ``children'' of the mediators in the causal Bayesian network. See example in tutorial.}

\item{response_model_exposure}{A model fitted for the response in a causal Bayesian network excluding ``children'' of the exposure and risk factors in the causal Bayesian network. See example in tutorial.}

\item{in_out}{A list of length 2. The first list contains a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list conttains a list of a single name of exposure or risk factor or outcome in form of characters. See tutorial examples for examples.}

\item{model_list}{By default this is set to an empty list. In the default setting, the models are fitted based on the order of the variables input in the parameter in_out. See the tutorial for more examples. Alternatively, the user can supply their own fitted models here by populating ``model_list'' with their own fitted models for each risk factor, mediator, exposure and response varialble. But the order of these models must be in the same order of the variables in the second list of in_out. See tutorial for further examples.}

\item{w}{Column of weights for case control matching listing in same order as patients in dataframe.}

\item{NumBootstrap}{The number of bootstraps the user wants to use to calculate confidence intervals for the effect. A minimum of 200 bootstrap repilcations (Efron (2016), Computer Age Statistical Inference, page 162) are recommended to calculate standard errors (for intervals of the form: estimate +/-1.96*(standard error of boostrap estimate. However increasing the number of bootstraps can result in the package taking a long time to run. So the user may make to balance speed with accuracy depedning on which is of more value in context.}

\item{NumSimulation}{This is the number of simulatons requested by the user to estimate integrals. The larger the number of simulations the more accurate the results but the longer the code takes to run. Therefore the user may wish to balance speed with accuracy depedning on which is of more value in the specific context of interest. The integrals for continuous variables are estimated using simulation methods.}
}
\value{

}
\description{
Evaluates Total PAF, Direct PAF, Indirect PAF and Path Specific PAF for a user inputted number of bootstraps and integral simulations
}
\examples{
\dontrun{
# I don't want you to run this
}
in_vars = c("subeduc","moteduc","fatduc")
outvar = c("phys")
make_formula(in_vars,outvar)
}
\keyword{Attributable}
\keyword{Fraction}
\keyword{Population}
\keyword{Regression}
\keyword{models}
