% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMarkovDAG.R
\name{checkMarkovDAG}
\alias{checkMarkovDAG}
\title{Checks if the causal DAG satisfies the Markov condition}
\usage{
checkMarkovDAG(in_out)
}
\arguments{
\item{in_out}{A list of length 2. The first list contains a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list conttains a list of a single name of exposure or risk factor or outcome in form of characters. See tutorial examples for examples.}
}
\value{
\item{IsMarkovDAG }{Returns a logical TRUE or FALSE whether it is a Markov DAG provided in_out is input as described in the documentation.}
\item{in_out}{The in_out list supplied in the function is returns the same of the input if IsMarkovDAG is returned TRUE. If IsMarkovDAG is returned FALSE the order of the in_out list is updated such that all parent variables come before ancestors in both i_out[[1]] and in_out[[2]]. This corrects any error where variables from a given Markov Bayesiand DAG are input to the package in the incorrect order.}
}
\description{
The functions checks if the Markov condition holds for the Directed Acyclic Graph (DAG) defined. Sometimes called the Markov assumption, is an assumption made in Bayesian probability theory, that every node in a Bayesian network is conditionally independent of its nondescendents, given its parents. In other words, it is assumed that a node has no bearing on nodes which do not descend from it. This is equivalent to stating that a node is conditionally independent of the entire network, given its Markov blanket. The related Causal Markov condition states that, conditional on the set of all its direct causes, a node is independent of all variables which are not direct causes or direct effects of that node.
}
\examples{
\dontrun{
# I don't want you to run this
}
in_vars = c("subeduc","moteduc","fatduc")
outvar = c("phys")
make_formula(in_vars,outvar)
}
\keyword{Acyclic}
\keyword{Attributable}
\keyword{Bayesian}
\keyword{Directed}
\keyword{Fraction}
\keyword{Graph}
\keyword{Markov}
\keyword{Population}
\keyword{models}
