% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_PAF.R
\name{sequential_PAF}
\alias{sequential_PAF}
\title{Sequential Population Attributable Fraction in a Bayesian Network.}
\usage{
sequential_PAF(
  dataframe,
  model_list_var,
  weights = 1,
  in_outDAG,
  response,
  NumOrderRiskFactors,
  addCustom = FALSE,
  custom = ""
)
}
\arguments{
\item{dataframe}{A wide format dataframe containing all the risk factors, confounders, exposures and outcomes within the causal DAG Bayesian network.}

\item{model_list_var}{is a list of models fitted for each of the variables in in_out$outlist based on its parents given in in_out$inlist. By default this is set to an empty list. In the default setting, the models are fitted based on the order of the variables input in the parameter in_outArg. See the tutorial for more examples. Alternatively, the user can supply their own fitted models here by populating ``model_listArg'' with their own fitted models for each risk factor, mediator, exposure and response varialble. But the order of these models must be in the same order of the variables in the second list of in_outArg. See tutorial for further examples.}

\item{weights}{Column of weights for case control matching listed in the same order as the patients in the data e.g. from tutorial weights = strokedata$weights.}

\item{in_outDAG}{A list of length 2. The first list contains a list of character vectors of the parents of the exposure or risk factor or outcome which are either causes or confounders of the exposure or risk factor or outcome. The second list contains a list of a single name of exposure or risk factor or outcome in form of characters. See tutorial examples for examples.}

\item{response}{The name of the response column variable within dataframe in text format e.g. "case". The cases should be coded as 1 and the controls as 0.}

\item{NumOrderRiskFactors}{is the number of randomly sampled elimination orders (or random permutations of all the risk factors) computing Monte-Carlo sequential attributable fractions for each random permutation.}

\item{addCustom}{Logical TRUE or FALSE indicating whether a customised interaction term is to be added to the each regression. The interaction term can include splines.}

\item{custom}{text containing the customised interaction term to be added to each regression. The text should be enclosed in inverted commas. Splines can be included within the interactin terms. See tutorial for examples.}
}
\value{
\item{plot }{ Returns a plot showing the estimated sequential attributable fractions, by position in elmination order. 95 percent confidence intervals are plotted so that we can be 95 percent confident the true estimate (that would be calculated from the procedure when the number of simulations m approaches infinity) lies in the Monte Carlo interval around the point estimate. The estimates shaded red correspond to a Bayesian network with indirect effects, whereas the estimates shaded blue correspond to the Bayesian network with no indirect effects modelled. Note that the Monte Carlo error at position k incorporates variation due to random selection of the set of risk factors/exposures that are intervened on in stages 1,...k - 1, and also variation based on the recursive simulation of the disease response. }
\item{SAF_summary }{Returns the data used for the plot containing both the Bayesian network (labelled Bayesian network) with indirect effects modelled and a model (labelled usual) with no indirect effects modelled.}
}
\description{
Calculates and plots sequential population attributbale fractions (PAF) under a Bayesian network struture.
}
\examples{
\dontrun{
# I don't want you to run this
}
}
\keyword{(PAF)}
\keyword{attributable}
\keyword{fraction}
\keyword{population}
\keyword{sequential}
